client/requirements.txt
altgraph==0.17.4
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.3
MouseInfo==0.1.3
packaging==23.2
pefile==2023.2.7
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
pyinstaller==6.0.0
pyinstaller-hooks-contrib==2023.9
PyMsgBox==1.0.9
pynput==1.7.6
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
python-decouple==3.8
pytweening==1.0.7
pywin32-ctypes==0.2.2
requests==2.31.0
schedule==1.2.1
setuptools==69.0.3
six==1.16.0
tim==0.5
urllib3==2.1.0
Werkzeug==3.0.1
